fix: Use WordPress' native functions to make remote calls
Do not use self-built curl requests, but use WordPress' native
HTTP API to make calls to remote servers.

This solves part of the issues described in #3. It does not yet replace
the calls to `copy` which have been commented out in this diff.

Maybe WordPress'
[download_url](https://developer.wordpress.org/reference/functions/download_url/)
would be a suitable replacement here.

Ref #3